refactor(applets): split modutils and halt behind explicit seams
Behavior-preserving restructurings; output, exit codes, messages, and all
existing tests are unchanged.

- modutils (#762): extract the shared module-name/file parsing and dependency
  planning (validate*, modinfoField, buildDeps, gatedError) into planner.go, and
  keep each applet's CLI surface in its own file (insmod.go, rmmod.go,
  modprobe.go, depmod.go). A new validateModuleNames helper folds the duplicated
  per-name validation loop; rmmod/modprobe delegate to it.
- halt (#763): separate the shutdown action planner and reboot executor
  (action, stop) into planner.go and the wtmp record encoding into wtmp.go,
  introducing a pure encodeWtmpRecord seam that writeWtmp now serializes through;
  halt.go keeps only the CLI wiring.

Adds focused tests for the new validateModuleNames and encodeWtmpRecord helpers.
